requirements.txt
discord.py
python-dotenv
pillow
psutil
requests
numpy
pympler
tweepy